requirements.txt
pyyaml
pysimplegui